Move PS1 generation to a separate file
I recently did some experimentation with my PS1 prompt, and I discovered
some interesting (not-ideal) things about the way I was previously
generating the PS1. Also, I wanted to move the PS1 generation to a
separate file to make it easier to break it up into functions without
polluting ~/.bashrc.

Previously, the color variables `CYAN`, etc were defined in ~/.bashrc.
This isn't ideal because it's actually setting those variables on every
bash session. This is unlikely to cause problems, but also unnecessary
and confusing. It would be better to store those in local variables that
don't pollute the environment. This also means less to evaluate each
time the prompt is generated - the PS1 has the actual values rather than
references to variables containing those values. Overall, everything is
less fragile this way.

There are no actual changes to the PS1 prompt here, but I'm hoping that
moving it to its own file will make things easier to work with and
change in the future.