Break PS1 into segments
No changes to the prompt, but it's now broken up into functions that are
easier to manage.

I tried several powerline-style prompts and didn't fall in love with any
of them. But something that they seem to get right is the ability to
customize and easily reorder segments of the PS1 prompt. This commit
moves prompt segments into their own functions so it's easier to
understand and easier to work with a small piece.