Allow secrets to be skipped
Sometimes, it is useful to install dotfiles without secrets. My primary
use-case for this is to apply my dotfiles inside a Docker container (or
VS Code devcontainer or GitHub Codespace) where I might not want my
secrets to be permanently stored. But this could also be useful in a
variety of other situations where I want my config but not my secrets.