Add sourceDir to Chezmoi Config
Adding `sourceDir =` to the top of the config file should allow Chezmoi
to be installed to places other than `~/.local/share/chezmoi` by passing
an option to `chezmoi init`.

This should allow us to install dotfiles to `~/dotfiles` if we prefer.

This should also allow us to use dotfiles more easily with VS Code
remote containers where dotfiles are pre-installed.

https://www.chezmoi.io/docs/how-to/#use-chezmoi-with-github-codespaces-visual-studio-codespaces-or-visual-studio-code-remote---containers