Prompt to Install Ubuntu Apps
In most cases, I want to automatically run my scripts to install ubuntu
applications. In some cases, however, I don't want these to be run. For
example, in a dev container, or on some non-permanent dotfiles
installation. Prompt before installing these, and save the config.

Along the way, change `store_secrets` to a boolean. It's better to
contain the string logic in one place, and `data` is allowed to be
boolean.

Soon, I plan to add additional logic to the config file to default all
of these for a non-interactive install (as in a dev container).