Make installer scripts x-platform
Change `install_ubuntu_apps` to `install_apps`. Some scripts should only
run on Ubuntu, we'll check `.chezmoi.os` on those. Some scripts can run
on either macOS or Linux, and we'll allow them to by *not* checking the
OS.

This clears a path for us to add macOS-only sections or scripts by
checking the OS there as well.